+++ app/Http/Controllers/Auth/RegisterController.php
@@ -90,7 +90,7 @@ class RegisterController extends Controller
                 }
 
                 $restricted = RestrictedNames::get();
-                if (in_array($value, $restricted)) {
+                if (in_array(strtolower($value), array_map('strtolower', $restricted))) {
                     return $fail('Username cannot be used.');
                 }
             },
@@ -115,7 +115,7 @@ class RegisterController extends Controller
             'name'     => 'nullable|string|max:'.config('pixelfed.max_name_length'),
             'username' => $usernameRules,
             'email'    => $emailRules,
-            'password' => 'required|string|min:12|confirmed',
+            'password' => 'required|string|min:'.config('pixelfed.min_password_length').'|confirmed',
         ];
 
         return Validator::make($data, $rules);