+++ app/Http/Controllers/FederationController.php
@@ -34,7 +34,8 @@ class FederationController extends Controller
     public function nodeinfoWellKnown()
     {
         abort_if(!config('federation.nodeinfo.enabled'), 404);
-        return response()->json(Nodeinfo::wellKnown());
+        return response()->json(Nodeinfo::wellKnown())
+            ->header('Access-Control-Allow-Origin','*');
     }
 
     public function nodeinfo()
@@ -62,7 +63,8 @@ class FederationController extends Controller
         }
         $webfinger = (new Webfinger($profile))->generate();
 
-        return response()->json($webfinger, 200, [], JSON_PRETTY_PRINT);
+        return response()->json($webfinger, 200, [], JSON_PRETTY_PRINT)
+            ->header('Access-Control-Allow-Origin','*');
     }
 
     public function hostMeta(Request $request)
@@ -103,6 +105,17 @@ class FederationController extends Controller
         $headers = $request->headers->all();
         $payload = $request->getContent();
         dispatch(new InboxValidator($username, $headers, $payload))->onQueue('high');
+        return;
+    }
+
+    public function sharedInbox(Request $request)
+    {
+        abort_if(!config('federation.activitypub.enabled'), 404);
+        abort_if(!config('federation.activitypub.sharedInbox'), 404);
+
+        $headers = $request->headers->all();
+        $payload = $request->getContent();
+        dispatch(new InboxWorker($headers, $payload))->onQueue('high');
         return;
     }