+++ app/Util/ActivityPub/Inbox.php
@@ -5,20 +5,25 @@ namespace App\Util\ActivityPub;
 use Cache, DB, Log, Purify, Redis, Validator;
 use App\{
     Activity,
+    DirectMessage,
     Follower,
     FollowRequest,
     Like,
     Notification,
+    Media,
     Profile,
     Status,
     StatusHashtag,
+    UserFilter
 };
 use Carbon\Carbon;
 use App\Util\ActivityPub\Helpers;
+use Illuminate\Support\Str;
 use App\Jobs\LikePipeline\LikePipeline;
 use App\Jobs\FollowPipeline\FollowPipeline;
 
 use App\Util\ActivityPub\Validator\Accept as AcceptValidator;
+use App\Util\ActivityPub\Validator\Add as AddValidator;
 use App\Util\ActivityPub\Validator\Announce as AnnounceValidator;
 use App\Util\ActivityPub\Validator\Follow as FollowValidator;
 use App\Util\ActivityPub\Validator\Like as LikeValidator;
@@ -57,6 +62,12 @@ class Inbox
     {
         $verb = (string) $this->payload['type'];
         switch ($verb) {
+
+            case 'Add':
+                if(AddValidator::validate($this->payload) == false) { return; }
+                $this->handleAddActivity();
+                break;
+
             case 'Create':
                 $this->handleCreateActivity();
                 break;
@@ -121,9 +132,27 @@ class Inbox
         return Helpers::profileFetch($actorUrl);
     }
 
+    public function handleAddActivity()
+    {
+        // stories ;)
+    }
+
     public function handleCreateActivity()
     {
         $activity = $this->payload['object'];
+        $actor = $this->actorFirstOrCreate($this->payload['actor']);
+        if(!$actor || $actor->domain == null) {
+            return;
+        }
+        $to = $activity['to'];
+        $cc = $activity['cc'];
+        if(count($to) == 1 && 
+            count($cc) == 0 && 
+            parse_url($to[0], PHP_URL_HOST) == config('pixelfed.domain.app')
+        ) {
+            $this->handleDirectMessage();
+            return;
+        }
         if(!$this->verifyNoteAttachment()) {
             return;
         }
@@ -170,10 +199,131 @@ class Inbox
         return;
     }
 
+    public function handleDirectMessage()
+    {
+        $activity = $this->payload['object'];
+        $actor = $this->actorFirstOrCreate($this->payload['actor']);
+        $profile = Profile::whereNull('domain')
+            ->whereUsername(array_last(explode('/', $activity['to'][0])))
+            ->firstOrFail();
+
+        if(in_array($actor->id, $profile->blockedIds()->toArray())) {
+            return;
+        }
+
+        $msg = $activity['content'];
+        $msgText = strip_tags($activity['content']);
+
+        if(Str::startsWith($msgText, '@' . $profile->username)) {
+            $len = strlen('@' . $profile->username);
+            $msgText = substr($msgText, $len + 1);
+        }
+
+        if($profile->user->settings->public_dm == false || $profile->is_private) {
+            if($profile->follows($actor) == true) {
+                $hidden = false;
+            } else {
+                $hidden = true;
+            }
+        } else {
+            $hidden = false;
+        }
+
+        $status = new Status;
+        $status->profile_id = $actor->id;
+        $status->caption = $msgText;
+        $status->rendered = $msg;
+        $status->visibility = 'direct';
+        $status->scope = 'direct';
+        $status->url = $activity['id'];
+        $status->in_reply_to_profile_id = $profile->id;
+        $status->save();
+
+        $dm = new DirectMessage;
+        $dm->to_id = $profile->id;
+        $dm->from_id = $actor->id;
+        $dm->status_id = $status->id;
+        $dm->is_hidden = $hidden;
+        $dm->type = 'text';
+        $dm->save();
+
+        if(count($activity['attachment'])) {
+            $photos = 0;
+            $videos = 0;
+            $allowed = explode(',', config('pixelfed.media_types'));
+            $activity['attachment'] = array_slice($activity['attachment'], 0, config('pixelfed.max_album_length'));
+            foreach($activity['attachment'] as $a) {
+                $type = $a['mediaType'];
+                $url = $a['url'];
+                $valid = Helpers::validateUrl($url);
+                if(in_array($type, $allowed) == false || $valid == false) {
+                    continue;
+                }
+
+                $media = new Media();
+                $media->remote_media = true;
+                $media->status_id = $status->id;
+                $media->profile_id = $status->profile_id;
+                $media->user_id = null;
+                $media->media_path = $url;
+                $media->remote_url = $url;
+                $media->mime = $type;
+                $media->save();
+                if(explode('/', $type)[0] == 'image') {
+                    $photos = $photos + 1;
+                }
+                if(explode('/', $type)[0] == 'video') {
+                    $videos = $videos + 1;
+                }
+            }
+
+            if($photos && $videos == 0) {
+                $dm->type = $photos == 1 ? 'photo' : 'photos';
+                $dm->save();
+            }
+            if($videos && $photos == 0) {
+                $dm->type = $videos == 1 ? 'video' : 'videos';
+                $dm->save();
+            }
+        }
+
+        if(filter_var($msgText, FILTER_VALIDATE_URL)) {
+            if(Helpers::validateUrl($msgText)) {
+                $dm->type = 'link';
+                $dm->meta = [
+                    'domain' => parse_url($msgText, PHP_URL_HOST),
+                    'local' => parse_url($msgText, PHP_URL_HOST) == 
+                        parse_url(config('app.url'), PHP_URL_HOST)
+                ];
+                $dm->save();
+            }
+        }
+
+        $nf = UserFilter::whereUserId($profile->id)
+            ->whereFilterableId($actor->id)
+            ->whereFilterableType('App\Profile')
+            ->whereFilterType('dm.mute')
+            ->exists();
+
+        if($profile->domain == null && $hidden == false && !$nf) {
+            $notification = new Notification();
+            $notification->profile_id = $profile->id;
+            $notification->actor_id = $actor->id;
+            $notification->action = 'dm';
+            $notification->message = $dm->toText();
+            $notification->rendered = $dm->toHtml();
+            $notification->item_id = $dm->id;
+            $notification->item_type = "App\DirectMessage";
+            $notification->save();
+        }
+
+        return;
+    }
+
     public function handleFollowActivity()
     {
         $actor = $this->actorFirstOrCreate($this->payload['actor']);
-        $target = $this->profile;
+        $target = $this->actorFirstOrCreate($this->payload['object']);
         if(!$actor || $actor->domain == null || $target->domain !== null) {
             return;
         }
@@ -305,57 +455,71 @@ class Inbox
         }
         $actor = $this->payload['actor'];
         $obj = $this->payload['object'];
-        if(is_string($obj) == true) {
-            return;
-        }
-        $type = $this->payload['object']['type'];
-        $typeCheck = in_array($type, ['Person', 'Tombstone']);
-        if(!Helpers::validateUrl($actor) || !Helpers::validateUrl($obj['id']) || !$typeCheck) {
-            return;
-        }
-        if(parse_url($obj['id'], PHP_URL_HOST) !== parse_url($actor, PHP_URL_HOST)) {
-            return;
-        }
-        $id = $this->payload['object']['id'];
-        switch ($type) {
-            case 'Person':
-                    // todo: fix race condition
-                    return; 
-                    $profile = Helpers::profileFetch($actor);
-                    if(!$profile || $profile->private_key != null) {
-                        return;
-                    }
-                    Notification::whereActorId($profile->id)->delete();
-                    $profile->avatar()->delete();
-                    $profile->followers()->delete();
-                    $profile->following()->delete();
-                    $profile->likes()->delete();
-                    $profile->media()->delete();
-                    $profile->statuses()->delete();
-                    $profile->delete();
+        if(is_string($obj) == true && $actor == $obj && Helpers::validateUrl($obj)) {
+            $profile = Profile::whereRemoteUrl($obj)->first();
+            if(!$profile || $profile->private_key != null) {
                 return;
-                break;
-
-            case 'Tombstone':
-                    $profile = Helpers::profileFetch($actor);
-                    $status = Status::whereProfileId($profile->id)
-                        ->whereUri($id)
-                        ->orWhere('url', $id)
-                        ->orWhere('object_url', $id)
-                        ->first();
-                    if(!$status) {
-                        return;
-                    }
-                    $status->media()->delete();
-                    $status->likes()->delete();
-                    $status->shares()->delete();
-                    $status->delete();
+            }
+            Notification::whereActorId($profile->id)->delete();
+            $profile->avatar()->delete();
+            $profile->followers()->delete();
+            $profile->following()->delete();
+            $profile->likes()->delete();
+            $profile->media()->delete();
+            $profile->hashtags()->delete();
+            $profile->statuses()->delete();
+            $profile->delete();
+            return;
+        } else {
+            $type = $this->payload['object']['type'];
+            $typeCheck = in_array($type, ['Person', 'Tombstone']);
+            if(!Helpers::validateUrl($actor) || !Helpers::validateUrl($obj['id']) || !$typeCheck) {
+                return;
+            }
+            if(parse_url($obj['id'], PHP_URL_HOST) !== parse_url($actor, PHP_URL_HOST)) {
+                return;
+            }
+            $id = $this->payload['object']['id'];
+            switch ($type) {
+                case 'Person':
+                        $profile = Profile::whereRemoteUrl($actor)->first();
+                        if(!$profile || $profile->private_key != null) {
+                            return;
+                        }
+                        Notification::whereActorId($profile->id)->delete();
+                        $profile->avatar()->delete();
+                        $profile->followers()->delete();
+                        $profile->following()->delete();
+                        $profile->likes()->delete();
+                        $profile->media()->delete();
+                        $profile->hashtags()->delete();
+                        $profile->statuses()->delete();
+                        $profile->delete();
                     return;
-                break;
-            
-            default:
-                return;
-                break;
+                    break;
+
+                case 'Tombstone':
+                        $profile = Helpers::profileFetch($actor);
+                        $status = Status::whereProfileId($profile->id)
+                            ->whereUri($id)
+                            ->orWhere('url', $id)
+                            ->orWhere('object_url', $id)
+                            ->first();
+                        if(!$status) {
+                            return;
+                        }
+                        $status->directMessage()->delete();
+                        $status->media()->delete();
+                        $status->likes()->delete();
+                        $status->shares()->delete();
+                        $status->delete();
+                        return;
+                    break;
+                
+                default:
+                    return;
+                    break;
+            }
         }
     }